Time: 7 ms (10.77%), Space: 56.1 MB (18.06%) - LeetHub

+++ 0795-number-of-subarrays-with-bounded-maximum/0795-number-of-subarrays-with-bounded-maximum.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int count(vector<int>& nums, int k){
+    long long count(vector<int>& nums, int k){
     long long c = 0;
     int len = 0;
